implement add item cart functionality
write addItem() reducer function that adds to the quantity of an item already in the cart, or adds the item to the cart if not already present.

import useDispatch and addItem cart action to ProductDetail page. Dispatch the action in handleAddToCart function with the constructed item to be added to cart as the payload