fix properties and updating logic
Add totalQuantity and showCart properties to state. Change cart function to be two separate function to open and close cart. Add logic to update the quantity of the cart and the total price of the items in the cart.

Add deleteItem reducer function to have the abililty to delete the item completely from the cart view.